refactor: move Matomo, Askem and HDS-style handling of the BaseApp to new files
HCRC-106
